+++ components/Home/index.jsx
@@ -1,10 +1,9 @@
 import { useState, useContext } from 'react';
 import { ethers } from 'ethers';
 import PropTypes from 'prop-types';
-import Image from 'next/image';
 import { connect } from 'react-redux';
+import { notification } from 'antd/lib';
 import { get } from 'lodash';
-import { Alert } from 'antd';
 import { CustomButton } from 'common-util/Button';
 import { getBalance } from 'common-util/functions';
 import {
@@ -12,10 +11,10 @@ import {
   setErrorMessage as setErrorMessageFn,
 } from 'store/setup/actions';
 import { DataContext } from 'common-util/context';
-import { Ellipsis } from 'components/GlobalStyles';
-import { claimBalances } from './utils';
-import { getUrl, getToken } from './helpers';
-import { Container, MiddleContent, TransactionSuccessMessage } from './styles';
+import { getSaleContract } from 'common-util/Contracts';
+import AlertInfo from 'components/AlertInfo';
+import { COLOR } from 'util/theme';
+import { Container, MiddleContent } from './styles';
 
 const CONNECT_WALLET_MESSAGE = 'To see balances and claim them, connect wallet';
 const TRANSACTION_STATE = {
@@ -23,10 +22,23 @@ const TRANSACTION_STATE = {
   failure: 'failure',
 };
 
+export const getToken = ({ tokenName, token }) => (
+  <div className={`section ${tokenName}-section`}>
+    <div className="info">
+      <span className="token-name">{tokenName}</span>
+      <span className="balance">{token || '--'}</span>
+    </div>
+  </div>
+);
+
 const Home = ({
   account, chainId, setUserBalance, setErrorMessage,
 }) => {
-  const { web3Provider, olasBalances: tokens } = useContext(DataContext);
+  const {
+    web3Provider,
+    olasBalances: tokens,
+    setOlasBalances,
+  } = useContext(DataContext);
 
   const [isClaimLoading, setClaimLoading] = useState(false);
   const [transactionState, setTransactionState] = useState(null);
@@ -43,25 +55,49 @@ const Home = ({
 
   const handleClaim = async () => {
     setClaimLoading(true);
-    try {
-      const response = await claimBalances(account, web3Provider, chainId);
 
-      // if claim is successfully done, transition to SUCCESS state!
-      setTransactionState(TRANSACTION_STATE.success);
-      setTransactionId(get(response, 'transactionHash') || null);
+    const contract = getSaleContract(window.MODAL_PROVIDER, chainId);
 
-      /* re-fetch tokens, balance after 3 seconds */
-      setTimeout(async () => {
-        await setBalance(account);
-      }, 3000);
-    } catch (error) {
-      console.error(error);
+    contract.methods
+      .claim()
+      .send({ from: account })
+      .then((response) => {
+        // if claim is successfully done, transition to SUCCESS state!
+        setTransactionState(TRANSACTION_STATE.success);
+        setTransactionId(get(response, 'transactionHash') || null);
 
-      // if claim is unsuccessful, transition to FAILURE state!
-      setTransactionState(TRANSACTION_STATE.failure);
-    } finally {
-      setClaimLoading(false);
-    }
+        // SHOW SUCCESS NOTIFICATION
+        notification.success({
+          description: 'Transaction Successful',
+          style: { border: `1px solid ${COLOR.PRIMARY}` },
+        });
+
+        /* re-fetch tokens, balance after 3 seconds */
+        setTimeout(async () => {
+          await setBalance(account);
+
+          const balancesAfterClaim = await contract.methods
+            .claimableBalances(account)
+            .call();
+
+          setOlasBalances(balancesAfterClaim);
+        }, 3000);
+      })
+      .catch((error) => {
+        console.error(error);
+
+        // SHOW ERROR NOTIFICATION
+        notification.error({
+          description: 'Some error occured',
+          style: { border: `1px solid ${COLOR.RED}` },
+        });
+
+        // if claim is unsuccessful, transition to FAILURE state!
+        setTransactionState(TRANSACTION_STATE.failure);
+      })
+      .finally(() => {
+        setClaimLoading(false);
+      });
   };
 
   const bToken = get(tokens, 'buBalance');
@@ -110,45 +146,11 @@ const Home = ({
 
       <br />
       {account && (
-        <>
-          {transactionState === TRANSACTION_STATE.success && (
-            <Alert
-              type="info"
-              message={(
-                <TransactionSuccessMessage>
-                  <div>Transaction Submitted</div>
-                  <div className="t-id">
-                    Track on Etherscan:&nbsp;
-                    <a
-                      href={getUrl(chainId, transactionId)}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      aria-label="ropsten-transaction"
-                    >
-                      <Ellipsis>{transactionId}</Ellipsis>
-                      <span className="external-link">
-                        <Image
-                          src="/images/external-link.svg"
-                          alt="Transaction link"
-                          width={18}
-                          height={16}
-                        />
-                      </span>
-                    </a>
-                  </div>
-                </TransactionSuccessMessage>
-              )}
-            />
-          )}
-
-          {transactionState === TRANSACTION_STATE.failure && (
-            <Alert
-              message="Claim transaction failed – try again"
-              type="error"
-              showIcon
-            />
-          )}
-        </>
+        <AlertInfo
+          transactionState={transactionState}
+          chainId={chainId}
+          transactionId={transactionId}
+        />
       )}
     </Container>
   );
