+++ util/constants.jsx
@@ -1,1 +1,6 @@
 export const CHAIN_ID = [1, 5, 31337];
+
+export const TRANSACTION_STATE = {
+  success: 'success',
+  failure: 'failure',
+};
